backend/requirements.txt
azure-functions
azure-storage-blob
azure-data-tables
pandas
bcrypt
PyJWT
requests
